Split workspace pull command from projects engine
